Making combogen run + edits
Fixed code for DXM, seems to have been broken in the main branch somehow
Fixed code for pulling the live combos.json data. Not sure if this was only broken on my machine or was broken on the live version due to being so old?
Edited dimensions to be a more standard ratio of 21:9
Swapped dead QR codes for current ones, including the newer app
Plus a few edits to make this generator run on my windows machine